DVT-#166 데둥여지도 필터링 기능 중 API 명세가 확정되지 않은 부분 제거
