+++ src/components/UsersMap/UsersMap.tsx
@@ -1,18 +1,15 @@
 import { Map } from "react-kakao-maps-sdk";
-import { courses, generations, roles } from "../../../fixtures/selectDatas";
-import { common } from "../../constants";
-import { Formik } from "../../hooks/useUsersLocationsFormik";
+import { useQueryClient } from "react-query";
+import { courses, generations } from "../../../fixtures/selectDatas";
+import { common, COORDS, CourseKeyType } from "../../constants";
+import useUsersLocationsFormik from "../../hooks/useUsersLocationsFormik";
+import useUsersLocationsQuery from "../../hooks/useUsersLocationsQuery";
 import { Position } from "../../types/commonTypes";
-import { ResponseUserLocation } from "../../types/userLocation";
 import { getUserMarkerOverlays } from "../../utils/map/overlay";
-import Input from "../base/Input";
-import Text from "../base/Text";
 import { Container, MapFloatContainer } from "../MapgakcoMap/styles";
 import UserMarker from "../MapgakcoMap/UserMarker";
 import { UserData } from "../MyProfile/types";
 import {
-  Button,
-  ButtonWrapper,
   InputWrapper,
   SearchBarFormContainer,
   Select,
@@ -20,12 +17,37 @@ import {
 
 interface Props {
   center: Position;
-  usersLocations?: ResponseUserLocation[];
   currentUser: UserData;
-  formik: Formik;
 }
 
-const UsersMap = ({ center, usersLocations, currentUser, formik }: Props) => {
+const UsersMap = ({ center, currentUser }: Props) => {
+  const queryClient = useQueryClient();
+
+  const formik = useUsersLocationsFormik({
+    initialValues: {
+      name: "",
+      course: currentUser?.user?.course,
+      generation: currentUser?.user?.generation,
+      role: currentUser?.user?.role,
+    },
+    submitHandler: () => {
+      queryClient.invalidateQueries(["usersLocations", formik.values]);
+    },
+  });
+
+  const { data: usersLocations } = useUsersLocationsQuery({
+    // TODO: 빠른 개발을 위해 대한민국 전체 좌표 범위를 사용한다. 사용자가 지도의 범위를 수정하면 해당하는 좌표 범위만 보여주도록 하는 기능을 추후 도입한다.
+    course:
+      formik.values.course ||
+      (currentUser?.user?.course as CourseKeyType) ||
+      "FE",
+    generation: formik.values.generation || currentUser?.user?.generation,
+    currentNEY: COORDS.KOREA_NEY,
+    currentNEX: COORDS.KOREA_NEX,
+    currentSWY: COORDS.KOREA_SWY,
+    currentSWX: COORDS.KOREA_SWX,
+  });
+
   const userMarkerOverlays = getUserMarkerOverlays(usersLocations, currentUser);
 
   return (
@@ -61,7 +83,8 @@ const UsersMap = ({ center, usersLocations, currentUser, formik }: Props) => {
                 </option>
               ))}
             </Select>
-            <Select
+            {/* TODO: 명세가 확정되고 필요하면 역할별 검색 기능을 추가한다. */}
+            {/* <Select
               id="role"
               name="role"
               onChange={formik.handleChange}
@@ -74,24 +97,26 @@ const UsersMap = ({ center, usersLocations, currentUser, formik }: Props) => {
                   {label}
                 </option>
               ))}
-            </Select>
-            <Input
+            </Select> */}
+            {/* TODO: 명세가 확정되고 필요하면 이름 검색 기능을 추가한다. */}
+            {/* <Input
               type="text"
               name="name"
               placeholder={common.message.ENTER_SEARCH_TERM}
               onChange={formik.handleChange}
               onBlur={formik.handleBlur}
               value={formik.values.name}
-            />
+            /> */}
           </InputWrapper>
 
-          <ButtonWrapper>
+          {/* TODO: 명세가 확정되고 필요하면 이름 검색 기능을 추가한다. */}
+          {/* <ButtonWrapper>
             <Button type="submit">
               <Text size={12} color="white" strong ellipsisLineClamp={1}>
                 {common.text.SEARCH}
               </Text>
             </Button>
-          </ButtonWrapper>
+          </ButtonWrapper> */}
         </SearchBarFormContainer>
       </MapFloatContainer>
       <Map
